Options have been moved to common directory
