+++ sig/lib/servactory/inputs/input.rbs
@@ -1,7 +1,7 @@
 module Servactory
   module Inputs
     class Input
-      @collection_of_options: OptionsCollection
+      @collection_of_options: Maintenance::Attributes::OptionsCollection
 
       attr_reader name: Symbol
       attr_reader internal_name: Symbol?
@@ -24,23 +24,23 @@ module Servactory
 
       def add_basic_options_with: (type: (Class | Array[Class]), options: untyped) -> void
 
-      def add_required_option_with: (untyped options) -> OptionsCollection
+      def add_required_option_with: (untyped options) -> Maintenance::Attributes::OptionsCollection
 
-      def add_types_option_with: ((Class | Array[Class]) type) -> OptionsCollection
+      def add_types_option_with: ((Class | Array[Class]) type) -> Maintenance::Attributes::OptionsCollection
 
-      def add_default_option_with: (untyped options) -> OptionsCollection
+      def add_default_option_with: (untyped options) -> Maintenance::Attributes::OptionsCollection
 
-      def add_collection_option_with: ((Class | Array[Class]) type, untyped options) -> OptionsCollection
+      def add_collection_option_with: ((Class | Array[Class]) type, untyped options) -> Maintenance::Attributes::OptionsCollection
 
-      def add_inclusion_option_with: (untyped options) -> OptionsCollection
+      def add_inclusion_option_with: (untyped options) -> Maintenance::Attributes::OptionsCollection
 
-      def add_must_option_with: (untyped options) -> OptionsCollection
+      def add_must_option_with: (untyped options) -> Maintenance::Attributes::OptionsCollection
 
-      def add_prepare_option_with: (untyped options) -> OptionsCollection
+      def add_prepare_option_with: (untyped options) -> Maintenance::Attributes::OptionsCollection
 
-      def add_internal_option_with: (untyped options) -> OptionsCollection
+      def add_internal_option_with: (untyped options) -> Maintenance::Attributes::OptionsCollection
 
-      def collection_of_options: -> OptionsCollection
+      def collection_of_options: -> Maintenance::Attributes::OptionsCollection
 
       def options_for_checks: () -> Hash[Symbol, untyped]
