--- sig/lib/servactory/inputs/options_collection.rbs
@@ -1,21 +1,0 @@
-module Servactory
-  module Inputs
-    class OptionsCollection
-      extend Forwardable
-
-      @collection: Set[Option]
-
-      def initialize: (*untyped) -> void
-
-      def names: () -> Array[Symbol]
-
-      def validation_classes: () -> Array[Class]
-
-      def defined_conflict_code: -> Symbol
-
-      def options_for_checks: -> Hash[Symbol, untyped]
-
-      def find_by: (name: Symbol) -> Option?
-    end
-  end
-end
