+++ sig/lib/servactory/maintenance/attributes/options_collection.rbs
@@ -1,0 +1,23 @@
+module Servactory
+  module Maintenance
+    module Attributes
+      class OptionsCollection
+        extend Forwardable
+
+        @collection: Set[Option]
+
+        def initialize: (*untyped) -> void
+
+        def names: () -> Array[Symbol]
+
+        def validation_classes: () -> Array[Class]
+
+        def defined_conflict_code: -> Symbol
+
+        def options_for_checks: -> Hash[Symbol, untyped]
+
+        def find_by: (name: Symbol) -> Option?
+      end
+    end
+  end
+end
